Fixes #3728 Apply the representation info when creating a module from a snapshot
Container icons and descriptions are not intrinsic to a PK-Sim structure.
They are stamped by MoBiExportTask, which visits the simulation and copies
IconName and Description from the representation info repository. That pass
only ran on the "Export simulation to MoBi" path, so a module recreated in
MoBi from its embedded PK-Sim snapshot came back without them and the
simulation tree rendered organs with no icon.

Lift the visitor into RepresentationInfoUpdater and apply it in
SnapshotExchange.CreateModule and CreateModuleAndExportInputs, so a recreated
module matches an exported one.

The updater starts the repository before visiting: ContainsInfoFor reads the
cache directly and would return false for every object in the CLI container,
silently skipping the whole pass.
